Revert "CRUD"
