+++ package.json
@@ -13,13 +13,12 @@
     "@reduxjs/toolkit": "^2.8.2",
     "@tailwindcss/line-clamp": "^0.4.4",
     "@tailwindcss/vite": "^4.1.7",
-    "axios": "1.12.0",
+    "axios": "^1.9.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-redux": "^9.2.0",
     "react-router": "^7.6.0",
     "react-router-dom": "^7.6.1",
-    "react-spring": "^10.0.3",
     "zustand": "^5.0.5"
   },
   "devDependencies": {
@@ -34,6 +33,6 @@
     "globals": "^16.0.0",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.1.7",
-    "vite": "7.1.11"
+    "vite": "^6.3.5"
   }
 }
