--- src/animations/springConfigs.js
@@ -1,6 +1,0 @@
-export const SPRING_CONFIGS = {
-  soft: { tension: 50, friction: 30 },
-  smooth: { tension: 120, friction: 20 },
-  bouncy: { tension: 200, friction: 10 },
-  slow: { tension: 40, friction: 40 },
-};
