+++ src/components/ui/inputFields/InputField.module.css
@@ -2,7 +2,7 @@
 @tailwind utilities;
 
 .userSurveyFormInput{
-    @apply bg-white w-full mt-1 px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-[var(--clr-secondary)] focus:border-transparent
+    @apply w-full mt-1 px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-[var(--clr-secondary)] focus:border-transparent
 }
 .userSurveyFormLabel{
     @apply flex flex-col items-center gap-x-2
