+++ src/store/usePostsStore.js
@@ -17,10 +17,10 @@ export const usePostsStore = create((set) => ({
   errorPost: null,
   errorNeighbors: null,
 
-  getPost: async (_id) => {
+  getPost: async (id) => {
     set({ errorPost: null, loadingPost: true });
     try {
-      const post = await postsService.fetchPost(_id);
+      const post = await postsService.fetchPost(id);
       set({ loadingPost: false, post: post });
     } catch (e) {
       set({
@@ -41,55 +41,15 @@ export const usePostsStore = create((set) => ({
       });
     }
   },
-  createPost: async (data) => {
-    set({ errorPost: null, loadingPost: true });
+  getNeighborPosts: async (id) => {
+    set({ errorNeighbors: null, loadingNeighbors: true });
     try {
-      const post = await postsService.fetchCreatePost(data);
-      if (post?.token) {
-        localStorage.setItem("token", post.token);
-      }
-      set({ loadingPost: false, post: post.article });
-      set((state) => ({ posts: [post.article, ...state.posts] }));
-
-      return post.article;
+      const neighborPosts = await postsService.fetchNeighborPosts(id);
+      set({ loadingNeighbors: false, neighborPosts: neighborPosts });
     } catch (e) {
       set({
-        errorPost: e?.message || ERRORS_MESSAGE.errorPosts,
-        loadingPost: false,
-      });
-    }
-  },
-  deletePost: async (_id) => {
-    set({ error: null, loading: true });
-    try {
-      await postsService.fetchDelete(_id);
-
-      set((state) => ({
-        loading: false,
-        posts: state.posts.filter((post) => post._id !== _id),
-      }));
-    } catch (e) {
-      console.error("Помилка при видаленні статті:", e);
-    }
-  },
-  updatePost: async (data, _id) => {
-    set({ error: null, loadingPost: true });
-    try {
-      const post = await postsService.fetchUpdatePost(_id, data);
-      if (post?.token) {
-        localStorage.setItem("token", post.token);
-      }
-      set({ loadingPost: false, post: post.article });
-      set((state) => ({
-        posts: state.posts.map((p) => (p._id === post._id ? post : p)),
-        loadingPost: false,
-      }));
-
-      return post.article;
-    } catch (e) {
-      set({
-        errorPost: e?.message || ERRORS_MESSAGE.errorPosts,
-        loadingPost: false,
+        errorNeighbors: e?.message || ERRORS_MESSAGE.errorPosts,
+        loadingNeighbors: false,
       });
     }
   },
