+++ src/store/useTotalStore.js
@@ -1,7 +1,8 @@
 import { create } from "zustand";
+import axios from "axios";
 
+import { SERVER_API_URL } from "@/constants/serviseApi.js";
 import { ERRORS_MESSAGE } from "@/constants/errorsConstants.js";
-import { postsService } from "@/service/postsService.js";
 
 export const useTotalStore = create((set) => ({
   total: 0,
@@ -11,8 +12,8 @@ export const useTotalStore = create((set) => ({
     set({ totalError: null, totalLoading: true });
 
     try {
-      const total = await postsService.fetchTotal();
-      return set({ total: total, totalLoading: false });
+      const response = await axios.get(`${SERVER_API_URL}/products?limit=1`);
+      return set({ total: response.data.total, totalLoading: false });
     } catch (error) {
       set({
         totalError: error?.message || ERRORS_MESSAGE.errorTotal,
